refactor: Phase 3 - split video.py into services/video/ (5 files)
- services/video/cache/frame_cache.py: VideoFrameCache (LRU + disk fallback)
- services/video/cache/legacy_cache.py: VideoCache (deprecated adapter)
- services/video/session.py: FFmpegSession (singleton + thread-safe)
- services/video/analyzer.py: VideoAnalyzer (scene detection + frame extract)
- services/video/processor.py: VideoProcessor (cut/concat/mix)
- video.py: compat layer with DeprecationWarning

VideoAnalyzer now uses static methods instead of class-level _ffmpeg_session.

Test: 384 passed, 20 skipped
